comment out contradictory collision resolution code
there was code to resolve collisions with other actors as if they were
safe to touch, as well as code to resolve collisions with actors causing
you to fly backwards. It lead to a race condition (I assume that
explains why the results of player/moblin collision varied based on the
presence or absence of some print statements. lol). Perhaps add a field
to each actor to specify whether or not they damage and knock back the
player?